requirements.txt
django
djangorestframework
mssql
mssql-django